6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float f;
-printf("size of a char: %lu byte(s)\n", (unsigned long)size of(a));
-printf("size of an int: %lu byte(s)\n"' (unsigned long)size of(b));
-printf("size of a long int: %lu byte(s)\n", (unsigned long)size of (c));
-printf("size of a long long int: %lu byte(s)\n",(unsigned long)size of (d));
-printf("size of a float: %lu byte(s)\n",(unsigned long)size of (f));
+printf("size of a char: %lu byte(s)\n", size of(a));
+printf("size of an int: %lu byte(s)\n"' size of(b));
+printf("size of a long int: %lu byte(s)\n", size of (c));
+printf("size of a long long int: %lu byte(s)\n", size of (d));
+printf("size of a float: %lu byte(s)\n", size of (f));
 return (0);
 }
